Reverted to previous behaviour when deleting links.
OK, I've had a bit of feedback concerning #90. While this simplifies the
logic, when clearing out a bunch of tags or from a search, it's not really
practical.
What this does is that it checks if we're on a page that needs to redirect
to self. so ?searchtags does, but ?Eox6ze (a permalink) does not.
Also, thanks to some magic, HTTP_REFERER is always set and therefore we
don't need to grab the current URL if redirecting to self.
Tested on every page with a delete link