Adds rel="noreferrer" to external links
A way to prevent Shaarli instance's URL to be "advertised" to other
sites is to use the HTML attribute "rel" of <a>. When set to
"noreferrer", it tells the browser not to send the referrer when
following the link.
Not all browser supports this but using it does not break those.

This commits sets the `rel` HTML attribute to `noreferrer`for all the
`<a>` that opens a bookmarked URL, no matter what referer
cleaner / redirector is configured.

Issue: #685